corrected first stop placement
