+++ SharedStops/SharedStopEnabler.csproj
@@ -70,13 +70,14 @@
     <Compile Include="RedirectionFramework\RedirectionUtil.cs" />
     <Compile Include="RedirectionFramework\Redirector.cs" />
     <Compile Include="RedirectionFramework\Tuple.cs" />
-    <Compile Include="RoadBridgeAIExt.cs" />
     <Compile Include="SharedStops.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StopSelection\AI\RoadBridgeAIExt.cs" />
     <Compile Include="StopSelection\Patch\NetSegmentPatch.cs" />
-    <Compile Include="StopSelection\StopsUtil.cs" />
-    <Compile Include="StopSelection\Patch\TransportLineAIPatch.cs" />
     <Compile Include="StopSelection\Patch\TransportLinePatch.cs" />
+    <Compile Include="StopSelection\StopsUtil.cs" />
+    <Compile Include="StopSelection\Patch\RoadAIPatch.cs" />
+    <Compile Include="StopSelection\Patch\TransportToolPatch.cs" />
     <Compile Include="StopSelection\SharedStopSegment.cs" />
     <Compile Include="StopSelection\SharedStopsTool.cs" />
     <Compile Include="StopSelection\StopSelectionExtensions.cs" />
@@ -91,6 +92,7 @@
       <Version>1.0.4</Version>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup Condition=" '$(OS)' == 'Windows_NT' ">
     <PostBuildEvent>mkdir "%25LOCALAPPDATA%25\Colossal Order\Cities_Skylines\Addons\Mods\$(SolutionName)"
