+++ SharedStops/StopSelection/Patch/RoadAIPatch.cs
